feat: sidebar/toolbar refactoring
