+++ frontend/src/widgets/Sidebar/model/types.ts
@@ -1,0 +1,6 @@
+import { TreeNode } from "@/entities/folder/model/types";
+
+export interface TreeNodeProps {
+  node: TreeNode;
+  level?: number;
+}
